eks/requirements.txt
pyyaml
Jinja2==2.11.1
MarkupSafe==1.1.1
boto3